+++ src/bittytax/conv/parsers/binance.py
@@ -419,6 +419,9 @@ def _parse_binance_statements_row(
         "Cash Voucher distribution",
         "Simple Earn Flexible Airdrop",
         "Campaign Related Reward",
+        "Launchpool Airdrop",
+        "HODLer Airdrops Distribution",
+        "Megadrop Rewards",
     ):
         data_row.t_record = TransactionOutRecord(
             TrType.AIRDROP,
@@ -429,6 +432,7 @@ def _parse_binance_statements_row(
         )
     elif row_dict["Operation"] in (
         "Distribution",
+        "Asset - Transfer",
         "Token Swap - Redenomination/Rebranding",
         "Token Swap - Distribution",
     ):
@@ -500,7 +504,53 @@ def _parse_binance_statements_row(
             sell_value=Decimal(0),
             wallet=WALLET,
         )
-    elif row_dict["Operation"] in ("Small assets exchange BNB", "Small Assets Exchange BNB"):
+    elif row_dict["Operation"] == "Binance Card Spending":
+        if Decimal(row_dict["Change"]) < 0:
+            data_row.t_record = TransactionOutRecord(
+                TrType.SPEND,
+                data_row.timestamp,
+                sell_quantity=abs(Decimal(row_dict["Change"])),
+                sell_asset=row_dict["Coin"],
+                wallet=WALLET,
+            )
+        else:
+            data_row.t_record = TransactionOutRecord(
+                TrType.FEE_REBATE,
+                data_row.timestamp,
+                buy_quantity=Decimal(row_dict["Change"]),
+                buy_asset=row_dict["Coin"],
+                wallet=WALLET,
+            )
+    elif row_dict["Operation"] == "Binance Card Cashback":
+        data_row.t_record = TransactionOutRecord(
+            TrType.CASHBACK,
+            data_row.timestamp,
+            buy_quantity=abs(Decimal(row_dict["Change"])),
+            buy_asset=row_dict["Coin"],
+            wallet=WALLET,
+        )
+    elif row_dict["Operation"] == "Crypto Box":
+        if Decimal(row_dict["Change"]) < 0:
+            data_row.t_record = TransactionOutRecord(
+                TrType.GIFT_SENT,
+                data_row.timestamp,
+                sell_quantity=abs(Decimal(row_dict["Change"])),
+                sell_asset=row_dict["Coin"],
+                wallet=WALLET,
+            )
+        else:
+            data_row.t_record = TransactionOutRecord(
+                TrType.GIFT_RECEIVED,
+                data_row.timestamp,
+                buy_quantity=Decimal(row_dict["Change"]),
+                buy_asset=row_dict["Coin"],
+                wallet=WALLET,
+            )
+    elif row_dict["Operation"] in (
+        "Small assets exchange BNB",
+        "Small Assets Exchange BNB",
+        "BNB Fee Deduction",
+    ):
         if config.binance_multi_bnb_split_even:
             _make_bnb_trade(
                 _get_op_rows(tx_times, data_row.timestamp, (row_dict["Operation"],)),
@@ -541,6 +591,7 @@ def _parse_binance_statements_row(
         "Transfer Between Spot Account and UM Futures Account",
         "Transfer Between Spot Account and CM Futures Account",
         "Transfer Between Main Account/Futures and Margin Account",
+        "Transfer Between Main and Funding Wallet",
         "Launchpool Subscription/Redemption",
         "Launchpad Subscribe",
         "Simple Earn Flexible Subscription",  # See merger
@@ -551,7 +602,11 @@ def _parse_binance_statements_row(
     ):
         # Skip non-taxable events and those which are handled by the merger
         return
-    elif row_dict["Operation"] in ("Deposit", "Fiat Deposit"):
+    elif row_dict["Operation"] in (
+        "Deposit",
+        "Fiat Deposit",
+        "Fiat OCBS - Add Fiat and Fees",
+    ):
         if config.binance_statements_only:
             data_row.t_record = TransactionOutRecord(
                 TrType.DEPOSIT,
@@ -563,15 +618,7 @@ def _parse_binance_statements_row(
         else:
             # Skip duplicate operations
             return
-    elif row_dict["Operation"] == "Send":
-        data_row.t_record = TransactionOutRecord(
-            TrType.WITHDRAWAL,
-            data_row.timestamp,
-            sell_quantity=abs(Decimal(row_dict["Change"])),
-            sell_asset=row_dict["Coin"],
-            wallet=WALLET,
-        )
-    elif row_dict["Operation"] in ("Withdraw", "Fiat Withdraw"):
+    elif row_dict["Operation"] in ("Withdraw", "Fiat Withdraw", "Fiat Withdrawal", "Send"):
         if config.binance_statements_only:
             data_row.t_record = TransactionOutRecord(
                 TrType.WITHDRAWAL,
@@ -583,7 +630,7 @@ def _parse_binance_statements_row(
         else:
             # Skip duplicate operations
             return
-    elif row_dict["Operation"] in ("Binance Convert", "Large OTC trading"):
+    elif row_dict["Operation"] in ("Binance Convert", "Large OTC trading", "Buy Crypto With Card"):
         if config.binance_statements_only:
             _make_trade(
                 _get_op_rows(tx_times, data_row.timestamp, (row_dict["Operation"],)),
@@ -659,7 +706,7 @@ def _parse_binance_statements_futures_row(
             sell_asset=row_dict["Coin"],
             wallet=WALLET,
         )
-    elif row_dict["Operation"] == "Funding Fee":
+    elif row_dict["Operation"] in ("Funding Fee", "Insurance Fund Refund"):
         if Decimal(row_dict["Change"]) > 0:
             data_row.t_record = TransactionOutRecord(
                 TrType.MARGIN_FEE_REBATE,
@@ -716,7 +763,11 @@ def _parse_binance_statements_margin_row(
             buy_asset=row_dict["Coin"],
             wallet=WALLET,
         )
-    elif row_dict["Operation"] in ("Margin Repayment", "Isolated Margin Repayment"):
+    elif row_dict["Operation"] in (
+        "Margin Repayment",
+        "Isolated Margin Repayment",
+        "Cross Margin Liquidation - Repayment",
+    ):
         data_row.t_record = TransactionOutRecord(
             TrType.LOAN_REPAYMENT,
             data_row.timestamp,
@@ -750,7 +801,19 @@ def _parse_binance_statements_margin_row(
                 ),
             ),
         )
-
+    elif row_dict["Operation"] in (
+        "Small assets exchange BNB",
+        "Small Assets Exchange BNB",
+        "BNB Fee Deduction",
+    ):
+        if config.binance_multi_bnb_split_even:
+            _make_bnb_trade(
+                _get_op_rows(tx_times, data_row.timestamp, (row_dict["Operation"],)),
+            )
+        else:
+            _make_trade(
+                _get_op_rows(tx_times, data_row.timestamp, (row_dict["Operation"],)),
+            )
     elif row_dict["Operation"] == "Transfer Between Main Account/Futures and Margin Account":
         # Skip not taxable events
         return